// agritechplatform/src/app/model/soil.model.ts
export interface SoilData {
    OBJECTID: number;
    Region: string;
    LANSLU: string;
    Soil_Subgroup: string;
    Shape_Area: number;
    Shape_Length: number;
    Primary_Soil_Classification: string;
    Primary_Soil_Value: number;
    Secondary_Soil_Classification: string;
    Secondary_Soil_Value: number;
    Tertiary_Soil_Classification: string;
    Tertiary_Soil_Value: number;
    Total_Soil_Classifications_Count: number;
    Total_Soil_Classification_Value: number;
}